"Update key features, configuration security and project metadata
This update includes several changes across the project. First, the feature 'User alignment with groups' has been marked as completed in README.md, indicating new functionality has been successfully added.

Environmental variables in 'handleCommands.js' script were corrected to use the recognized format. Sensitive data such as 'token' and 'clientId' have been moved to environment variables for improved security and code portability.

In 'groupManager.js', the requirement for the amount of money to create a group was corrected. The configuration values for the groupId were updated and consolidated to use environment variables, enhancing security.

Changes in 'ask.js' were done to better help with the generation of the context for the bot's question making it more clear and detailed. Improvements were done to error handling as well.

In 'ready.js', the message now shows total members rather than the specific "MWT GNs" members. To match the new, more general output, the bot's status message was updated.

The 'package.json' was added to include project metadata and manage dependencies. This makes it easier for others to understand and contribute to the project.

Files that should be ignored like logs, cache, or specific OS and IDE files were added to the '.gitignore'.

Updating the 'package-lock.json' and some 'index.js' values were corrected to more standard representations."