Add XP and Level to userModel and refine starboardMessageModel
This commit adds 'XP' and 'level' fields to the 'userModel' to support a new feature of user progress tracking. In addition, the 'starboardMessageModel' is updated to no longer have a 'sent' field but instead a 'starboardMessageID' field to better handle the process of tracking and storing messages that have been 'starstruck'.