Add ephemeral option to embed messages
Updated the "replyWithEmbed" function in "embedResponse.no.js". This allows for an optional parameter "ephemeral" to be passed which determines if the message should be private or not. In addition, the expected type of "actionRow" parameter has been clarified in the function's docstring. This was done to support more customization in how our bot communicates, and hence, improve the user experience.