Update group member check and version patch
The code in groupManager.js was updated to use the JavaScript includes() method to check if a user is part of a group, instead of the non-existing in operator. This is a more idiomatic way to perform this check in JavaScript and prevents runtime errors. The version number in package.json was also incremented to reflect this bug fix.