Update package versions and refactor group info command
The package versions have been updated in the "package-lock.json" and "package.json" files to use newer versions of the dependencies. In particular, the "axios" and "form-data" modules are removed, and the "semver" module is added for semantic versioning.

Also, the logic for fetching user details in the 'group' command has been updated to handle the case when the user is 'Vacant'. The embed footer has also been modified to use the requester's details. The "group list" command has been enhanced for user-friendly group name display.

These changes ensure the application dependencies are up-to-date and improve the user experience in the 'group' command.