"Added pm2 dependency and updated package-lock.json
Added the pm2 (Process Manager 2) dependency to manage application processes. This includes enhancements for handling log management, application monitoring, and cluster mode. This change reflects in an updated package-lock.json file.

The diff also shows an addition of numerous other dependencies which are mostly related to pm2 functionalities, as well as related modules and sub-dependencies. Also, updated versions of some dependencies are evident."